# swaggerDocumentation.yaml
openapi: 3.0.0
info:
  title: API Week-10 Version 1
  version: '1.0'
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /users:
    get:
      summary: Get All Users
      operationId: getAllUsers
      responses:
        '200':
          description: List of users
      security:
        - bearer: []
...
security:
  - securityRequirement: []
...